Refactor out cipher components from aead
